// frontend/src/api/generated/main-service/apiGenerated.ts
export interface TemplateResponse {
  activeVersion?: null | TemplateVersionResponse
  allVersions?: TemplateVersionResponse[]
  sandboxPayload?: null | string
  /** @format uuid */
  id: string
  /** @format uuid */
  userId: string
  /** @format uuid */
  projectId?: null | string
  templateKey: string
  name: string
  createdBy?: null | string
  /** @format date-time */
  createdDatetime?: any
  updatedBy?: null | string
  /** @format date-time */
  updatedDatetime?: any
}

export interface TemplateVersionResponse {
  previewFilePathPresigned?: null | string
  /** @format uuid */
  id: string
  /** @format uuid */
  templateId: string
  /**
   * @format int32
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  version: number | string
  filePath: string
  status?: null | string
  previewFilePath?: null | string
  sandboxPayload?: null | string
  sandboxFilePath?: null | string
  createdBy?: null | string
  /** @format date-time */
  createdDatetime?: any
  updatedBy?: null | string
  /** @format date-time */
  updatedDatetime?: any
}